Created repository for Emphatic, a FAT compatible file system that aims to avoid file fragmentation.
Started work on a program to mount an emphatic file system, using the Linux
FUSE system. Under this, the mount program basically becomes a daemon which
services file operation requests on the mounted file system. This file
(fuse.c) will probably be changed to a different name in a later commit.