Started work on implementing the procedures defined in fileio.h. So far I have completed fat_open.
Added a couple of short-cut macros to fat.h, for obtaining the first sector
in the FAT, and the sector size in bytes.

Added a size field to the fat_file_t structure.

Added a type definition for fat_entry_t, which could potentially be given
a different definition for FAT12/16? The downside of that is that it would
require recompiling.