Added an implementation for fat_write in fileio.c.
Added a new field current_cluster in fat_file_t, to hold a pointer to the
cluster list entry of the cluster where the next read or write will begin.
This prevents us from having to walk along the cluster list for every
read or write operation, and eliminates duplicated code.

Added a couple of utility functions to fileio.c. update_current_cluster
will reset the current cluster field of a file descriptor to point to the
cluster list entry of the cluster where the next read or write will begin,
and count_clusters returns the number of clusters required to accomodate
a given number of bytes.

Major work on fileio.c is now complete. Directory lookup routines, and
a cluster allocation routine still need to be implemented.