Started work on the directory lookup code, with directory.c and .h. I have started on a simple implementation for fat_lookup_dir, which recursively breaks the path down into a file to look up, and the parent directory where the file is to be found. Once it reaches the root directory, it will fill in a dummy directory structure based on the fields in the volume struct.
Although this lookup routine does not use any caching techniques to improve
its efficiency, such improvements are on the wishlist. A likely improvement
would be to implement a directory translation lookaside buffer, which could
be done with only minor modifications to the current procedure, by making
a hit on the TLB the base case of the recursion.

Other changes in this commit are minor modifications to other source files,
as some headers were not #include'd when they should have been.