Defined a bool type in const.h, and defined values for true and false.
Changed the tabulation (tabs are replaced by four space characters), and
added modelines to tell Vim to behave accordingly. This unfortunately has
resulted in *many* changes, so I do NOT want to have to do it again. If
you wish to contribute code, please make sure that you follow the above
convention.