Finished code for merging a freed cluster back into the pool of free clusters, and implemented get_nearest_free and get_largest_region. Also added a higher order function for traversing the free space list, which get_nearest and get_largest make use of, by providing a callback function which is invoked on every item of the free space list.
Note that these end up having to use pointer pointers quite a lot, because
get_nearest *may* have to unlink an item from the free space list (if the
nearest free region is only a single cluster).

Also fixed a typo in the Makefile; the build command line dependency was
for the source files, it should be for the object files.