More work on directory lookup code, in directory.c. Also added a volume_info global, and initialisation function. Added functions which are called from fat_lookup_dir. An ongoing issue is how we break up the full path name into a parent directory name, and file name: do we modify the string pointed to by fat_lookup_dir's path argument, or do we copy that string to a separate buffer first? At present, we copy to a separate buffer, but doing it in place would be more memory efficient.
Added a new macro to fat.h, which breaks up a 32 bit cluster address and
stores it in the upper and lower 16 bit fields of a directory entry.

Other modifications in this commit are mainly adding 'const' to pointer
params which should not be modified by the function concerned, like the
fat_volume_t * arguments to the *_init functions.