Added a handler for the utimens FUSE method. Also added calls to update file access and modified time stamps in the read and write handlers.
The functions which actually update the times are still to be implemented,
and are complicated by the fact that the time stamps are kept in the
directory entry, meaning that time stamp updates require writing on the
parent directory. I am considering implementing a second files list to
hold active directory descriptors, which avoids the overhead involved in
opening and closing them.

Modified the parent directory field of the fat_file_t structure, to be a
FAT entry, rather than a pointer to an in memory copy of the directory
entry.