Added a handler for the FUSE truncate method in mfatic-fuse.c
Modified the Makefile so that (hopefully) Depend will only be built when
we are compiling, not at other times eg. when makeing tags.

Added a field in the fat_file_t structure for the directory entry index,
which allows an open file's directory entry to be accessed in O(1) time.