Added new procedures exported from directory.c which allow directory entries to be accessed easily, by specifying a directory i-node, and entry index. Also added a global file_list_t * to track the currently active directories.
Added two procedures exported from dostimes.c to update the accessed and
modified time stamps. Both make use of the dir entry read/write functions.

Added a new field to the fat_file_t structure, to keep track of the number
of tables it is referenced in. This is necessary in order for ilist_unlink
to be able to know whether it has to release the structure or not.

Also modified the Makefile, because it was still building Depend when I ran
the tags target, which makes no sense, because tags does not depend on
Depend. Fixed by removing $(SRC) from Depend's dependencies.