Added code to handle create/delete operations.
Also modified dir_lookup_file to provide the file handle of the parent
directory, and not add the parent dir to the active directories list. This
will simplify system calls that use dir_lookup_file, but do not affect the
active directories list, such as stat.

Added a new flags field to the fat_file_t struct, which ensures correct
operation of the unlink system call (ie. files are only released when all
open instances are closed).