Renamed old fat_open (which opens using a directory entry).
Created a new fat_open with the same parameter list as used by several
existing calls to fat_open, which is a useful abstraction. It takes a
path name for a file to be opened, and creates a file structure.

The new fat_open actually is just a function that calls fat_lookup_dir
and then fat_open_fd (old fat_open) sequentially, thereby not causing
nasty mutual recursion (because fat_lookup_dir calls fat_open_fd to read
from directories to traverse a path name).